refactor(providers): extract shared scaffolding into llm-router; fix UTF-8 chunk splits
Every provider crate carried verbatim copies of the same plumbing —
router_client.rs was byte-identical across four crates, state.rs differed
by one constant, and the stream pump literally carried a 'shared
extraction into llm-router is a listed follow-up' comment in four crates.
The copies had already drifted: only provider-anthropic buffered UTF-8
across network chunks; the OpenAI-style providers decoded each chunk with
String::from_utf8_lossy, corrupting any multibyte codepoint split across
two chunks into U+FFFD.

New llm_router::provider_scaffold (providers already path-dep the lib):
- router_client: provider-protocol wrappers, parameterized by provider id
  (openai-codex keeps its vault/oauth extras locally, sharing 'call')
- state: registration-token persistence, parameterized by scope
- pump: send_event + heartbeat pump (tests moved with it)
- sse_transport: error_chain, append_utf8_chunk (cross-chunk UTF-8
  buffering with regression tests), normalize_crlf, and a generic
  drain_sse_blocks over a per-provider decode closure
- names: the ::/__ tool-name codec

All six providers now import these; the five OpenAI-style providers pick
up append_utf8_chunk (fixing the split-codepoint corruption) and CRLF
block reframing (SSE permits CRLF; bare find("\n\n") missed it).
Provider-specific SSE decoding, request building, and error
classification stay per-crate. llamacpp's error_chain copy lived in
errors.rs and is gone too; its integration test gained a bounded retry
around refresh_models for a pre-existing config-propagation race
(verified failing on the unconverted baseline).

Tests: cargo test green in llm-router and all six provider crates
(engine-backed integration suites included).
