+++ provider-openai-codex/src/stream_fn.rs
@@ -14,15 +14,11 @@ use iii_sdk::errors::Error;
 use iii_sdk::IIIClient;
 use llm_router::channels::open_sink;
 use llm_router::chat::relay::FrameSink;
-use llm_router::types::events::{AssistantMessageEvent, ErrorKind};
+use llm_router::provider_scaffold::pump::{pump, send_event, PING_INTERVAL};
+use llm_router::types::events::ErrorKind;
 use llm_router::types::router::{
     CredentialSource, ProviderResolveResponse, ProviderStreamInput, ProviderStreamOutput,
 };
-use std::time::Duration;
-use tokio::sync::mpsc;
-
-/// Heartbeat cadence while the upstream is silent (spec: at least every 30s).
-pub const PING_INTERVAL: Duration = Duration::from_secs(30);
 
 pub fn make_stream(
     iii: IIIClient,
@@ -40,11 +36,6 @@ pub fn make_stream(
             Ok(ProviderStreamOutput { ok: true })
         })
     }
-}
-
-fn send_event(sink: &dyn FrameSink, ev: &AssistantMessageEvent) -> Result<(), ()> {
-    let frame = serde_json::to_string(ev).expect("serializable event");
-    sink.send(&frame).map_err(|_| ())
 }
 
 fn default_resolve() -> ProviderResolveResponse {
@@ -150,76 +141,4 @@ async fn run_stream_call(
         },
     );
     pump(rx, sink, PING_INTERVAL).await;
-}
-
-/// Forward upstream events to the sink; ping through silence; stop on the
-/// terminal event or a failed write (caller gone → dropping `rx` aborts the
-/// upstream task and its in-flight HTTP request).
-pub async fn pump(
-    mut rx: mpsc::Receiver<AssistantMessageEvent>,
-    sink: &dyn FrameSink,
-    ping_interval: Duration,
-) {
-    loop {
-        match tokio::time::timeout(ping_interval, rx.recv()).await {
-            Ok(Some(ev)) => {
-                let terminal = ev.is_terminal();
-                if send_event(sink, &ev).is_err() {
-                    return;
-                }
-                if terminal {
-                    return;
-                }
-            }
-            Ok(None) => return,
-            Err(_elapsed) => {
-                if send_event(sink, &AssistantMessageEvent::Ping).is_err() {
-                    return;
-                }
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::sse::empty_assistant;
-    use llm_router::chat::relay::RelayRead;
-    use llm_router::testkit::fake_channels::FakeChannel;
-    use serde_json::Value;
-
-    fn done_event() -> AssistantMessageEvent {
-        AssistantMessageEvent::Done {
-            message: empty_assistant("codex/gpt-5.5"),
-        }
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn forwards_events_and_stops_at_terminal() {
-        let ch = FakeChannel::new();
-        let (tx, rx) = mpsc::channel(8);
-        tx.send(AssistantMessageEvent::Start {
-            partial: empty_assistant("m"),
-        })
-        .await
-        .unwrap();
-        tx.send(done_event()).await.unwrap();
-        tx.send(AssistantMessageEvent::Ping).await.unwrap();
-        drop(tx);
-
-        pump(rx, &ch.writer, Duration::from_secs(30)).await;
-        ch.writer.close();
-
-        let mut frames = Vec::new();
-        let mut reader = ch.reader;
-        while let llm_router::chat::relay::ReadEvent::Msg(m) =
-            reader.next(Duration::from_millis(100)).await
-        {
-            frames.push(m);
-        }
-        assert_eq!(frames.len(), 2);
-        let last: Value = serde_json::from_str(&frames[1]).unwrap();
-        assert_eq!(last["type"], "done");
-    }
 }
