+++ provider-zai/src/state.rs
@@ -1,34 +1,15 @@
-//! Registration-token persistence in iii-state (engine `state::*` functions,
-//! binary-worker.md § 7). The raw token lives here, under the provider's own
-//! scope; the router persists only its sha256 hash.
+//! Registration-token persistence, scoped to this provider's worker id
+//! (shared logic in `llm_router::provider_scaffold::state`).
 use iii_sdk::errors::Error;
-use iii_sdk::protocol::TriggerRequest;
 use iii_sdk::IIIClient;
-use serde_json::{json, Value};
+use llm_router::provider_scaffold::state as scaffold;
 
 pub const STATE_SCOPE: &str = "provider-zai";
-const TOKEN_KEY: &str = "registration_token";
 
 pub async fn load_token(iii: &IIIClient) -> Option<String> {
-    let value = iii
-        .trigger(TriggerRequest {
-            function_id: "state::get".into(),
-            payload: json!({ "scope": STATE_SCOPE, "key": TOKEN_KEY }),
-            action: None,
-            timeout_ms: None,
-        })
-        .await
-        .ok()?;
-    value.as_str().map(String::from)
+    scaffold::load_token(iii, STATE_SCOPE).await
 }
 
 pub async fn store_token(iii: &IIIClient, token: &str) -> Result<(), Error> {
-    iii.trigger(TriggerRequest {
-        function_id: "state::set".into(),
-        payload: json!({ "scope": STATE_SCOPE, "key": TOKEN_KEY, "value": Value::from(token) }),
-        action: None,
-        timeout_ms: None,
-    })
-    .await?;
-    Ok(())
+    scaffold::store_token(iii, STATE_SCOPE, token).await
 }
