+++ provider-zai/src/upstream.rs
@@ -5,6 +5,9 @@
 use crate::errors::classify;
 use crate::sse::{build_final, build_partial, handle_chunk, synthetic_error_event, PartialState};
 use futures::StreamExt;
+use llm_router::provider_scaffold::sse_transport::{
+    append_utf8_chunk, drain_sse_blocks, error_chain,
+};
 use llm_router::types::events::{AssistantMessageEvent, ErrorKind};
 use serde_json::Value;
 use tokio::sync::mpsc;
@@ -27,66 +30,6 @@ pub fn spawn_upstream(
         run_upstream(client, args, tx).await;
     });
     rx
-}
-
-/// Flatten an error and its `source()` chain into one string. reqwest's
-/// top-level Display for a builder error is just "builder error"; the real
-/// cause (invalid header value, bad URL) lives in the source chain, so without
-/// this the message is undiagnosable.
-fn error_chain(e: &dyn std::error::Error) -> String {
-    let mut msg = e.to_string();
-    let mut src = e.source();
-    while let Some(s) = src {
-        let next = s.to_string();
-        // reqwest sometimes nests the same text; skip exact repeats.
-        if !msg.ends_with(&next) {
-            msg.push_str(": ");
-            msg.push_str(&next);
-        }
-        src = s.source();
-    }
-    msg
-}
-
-/// Append chunk bytes to `text`, retaining any trailing incomplete UTF-8
-/// sequence: transport chunk boundaries are arbitrary, so a multi-byte
-/// character (GLM output is frequently CJK) can split across chunks — a
-/// per-chunk lossy decode would corrupt it into U+FFFD.
-fn append_utf8_chunk(byte_buf: &mut Vec<u8>, text: &mut String, chunk: &[u8]) {
-    byte_buf.extend_from_slice(chunk);
-    let mut consumed = 0usize;
-    loop {
-        match std::str::from_utf8(&byte_buf[consumed..]) {
-            Ok(s) => {
-                text.push_str(s);
-                byte_buf.clear();
-                return;
-            }
-            Err(e) => {
-                let valid = e.valid_up_to();
-                if valid > 0 {
-                    // SAFETY: valid_up_to guarantees valid UTF-8 in this prefix.
-                    text.push_str(unsafe {
-                        std::str::from_utf8_unchecked(&byte_buf[consumed..consumed + valid])
-                    });
-                    consumed += valid;
-                }
-                match e.error_len() {
-                    Some(invalid) => {
-                        byte_buf.drain(..consumed + invalid);
-                        text.push('\u{FFFD}');
-                        consumed = 0;
-                    }
-                    None => {
-                        if consumed > 0 {
-                            byte_buf.drain(..consumed);
-                        }
-                        return;
-                    }
-                }
-            }
-        }
-    }
 }
 
 /// Last `data: ` payload in an SSE block, if any.
@@ -148,7 +91,34 @@ async fn run_upstream(
 
     let mut stream = resp.bytes_stream();
     let mut buf = String::new();
-    let mut byte_buf: Vec<u8> = Vec::new();
+    // Cross-chunk UTF-8 buffering: transport chunk boundaries are arbitrary,
+    // so a multi-byte character (GLM output is frequently CJK) can split
+    // across chunks — a per-chunk lossy decode would corrupt it into U+FFFD.
+    let mut byte_buf = Vec::new();
+    // Block decoder: [DONE] closes the stream (Stop + Done, Done terminal);
+    // anything else parses and runs the chunk state machine.
+    let decode =
+        |data_block: &str, state: &mut PartialState, model: &str| -> Vec<AssistantMessageEvent> {
+            let Some(data) = data_line(data_block) else {
+                return vec![];
+            };
+            if data == "[DONE]" {
+                return vec![
+                    AssistantMessageEvent::Stop {
+                        stop_reason: state.stop_reason(),
+                        error_message: None,
+                        error_kind: None,
+                    },
+                    AssistantMessageEvent::Done {
+                        message: build_final(state, model),
+                    },
+                ];
+            }
+            let Ok(parsed) = serde_json::from_str::<Value>(data) else {
+                return vec![];
+            };
+            handle_chunk(&parsed, state, model)
+        };
     while let Some(chunk) = stream.next().await {
         let chunk = match chunk {
             Ok(c) => c,
@@ -164,38 +134,12 @@ async fn run_upstream(
             }
         };
         append_utf8_chunk(&mut byte_buf, &mut buf, &chunk);
-        while let Some(idx) = buf.find("\n\n") {
-            let block: String = buf.drain(..idx + 2).collect();
-            let Some(data) = data_line(&block) else {
-                continue;
-            };
-            if data == "[DONE]" {
-                let _ = tx
-                    .send(AssistantMessageEvent::Stop {
-                        stop_reason: state.stop_reason(),
-                        error_message: None,
-                        error_kind: None,
-                    })
-                    .await;
-                let _ = tx
-                    .send(AssistantMessageEvent::Done {
-                        message: build_final(&state, &args.model),
-                    })
-                    .await;
-                return;
-            }
-            let Ok(parsed) = serde_json::from_str::<Value>(data) else {
-                continue;
-            };
-            for ev in handle_chunk(&parsed, &mut state, &args.model) {
-                let terminal = ev.is_terminal();
-                if tx.send(ev).await.is_err() {
-                    return; // receiver dropped → abort upstream
-                }
-                if terminal {
-                    return; // exactly one terminal event
-                }
-            }
+        if drain_sse_blocks(&mut buf, &tx, &mut |block: &str| {
+            decode(block, &mut state, &args.model)
+        })
+        .await
+        {
+            return; // terminal forwarded, or receiver dropped → abort upstream
         }
     }
     // Stream ended without [DONE] (connection close framing): still terminal.
@@ -347,24 +291,6 @@ mod tests {
             }
             other => panic!("want done, got {other:?}"),
         }
-    }
-
-    #[test]
-    fn utf8_split_across_chunks_survives_reassembly() {
-        // "你" = 3 bytes; split it across two transport chunks.
-        let bytes = "data: 你好\n\n".as_bytes();
-        let (a, b) = bytes.split_at(8); // mid-character
-        let (mut byte_buf, mut text) = (Vec::new(), String::new());
-        append_utf8_chunk(&mut byte_buf, &mut text, a);
-        append_utf8_chunk(&mut byte_buf, &mut text, b);
-        assert_eq!(text, "data: 你好\n\n");
-        assert!(!text.contains('\u{FFFD}'), "lossy corruption: {text:?}");
-        assert!(byte_buf.is_empty());
-
-        // Truly invalid bytes still degrade to U+FFFD instead of stalling.
-        let (mut byte_buf, mut text) = (Vec::new(), String::new());
-        append_utf8_chunk(&mut byte_buf, &mut text, &[b'a', 0xFF, b'b']);
-        assert_eq!(text, "a\u{FFFD}b");
     }
 
     #[tokio::test(flavor = "multi_thread")]
