+++ provider-zai/src/wire/names.rs
@@ -1,28 +1,5 @@
-//! iii function ids ↔ Z.AI function names. Z.AI enforces
-//! `^[a-zA-Z0-9_-]{1,64}$`; bus ids use `::` separators.
+//! iii function ids ↔ Z.AI tool names. Z.AI enforces
+//! `^[a-zA-Z0-9_-]{1,64}$`; bus ids use `::` separators. Shared codec
+//! (and its tests) live in `llm_router::provider_scaffold::names`.
 
-pub fn encode_tool_name(name: &str) -> String {
-    name.replace("::", "__")
-}
-
-pub fn decode_tool_name(name: &str) -> String {
-    name.replace("__", "::")
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn round_trips_bus_ids() {
-        assert_eq!(encode_tool_name("web::fetch"), "web__fetch");
-        assert_eq!(decode_tool_name("web__fetch"), "web::fetch");
-        assert_eq!(decode_tool_name(&encode_tool_name("a::b::c")), "a::b::c");
-    }
-
-    #[test]
-    fn plain_names_pass_through() {
-        assert_eq!(encode_tool_name("submit_result"), "submit_result");
-        assert_eq!(decode_tool_name("submit_result"), "submit_result");
-    }
-}
+pub use llm_router::provider_scaffold::names::{decode_tool_name, encode_tool_name};
